Pruning & refactoring of get links functionality
